+++ docs_src/connections/force_rollback.py
@@ -1,0 +1,15 @@
+from databasez import Database
+
+database = Database("sqlite:///foo.sqlite", force_rollback=True)
+
+
+def test_foo():
+    async with database:
+        ...
+        # do the tests
+    # and now everything is rolled back
+
+    async with database:
+        ...
+        # do the tests
+    # and now everything is rolled back again