+++ docs_src/transactions/async_context_database.py
@@ -1,0 +1,12 @@
+from databasez import Database
+
+
+async def main():
+    async with Database("<URL>") as database:
+        # do something
+        async with database.transaction():
+            ...
+
+        # check something and then reset
+        async with database.transaction(force_rollback=True):
+            ...